refactor: move rect sample to vision package
